+++ GERTi/GERTiMNC.lua
@@ -1,4 +1,4 @@
--- GERT v1.5 Build 1
+-- GERT v1.5 Build 2
 local component = require("component")
 local computer = require("computer")
 local event = require("event")
@@ -19,18 +19,6 @@ local savedAddresses = {}
 local registeredPrograms = {} -- [programName] = {[1]=GERTiAddress1, [2]=GERTiAddress2, ...}
 local addressFile = "/etc/GERTaddresses.gert"
 local networkFile = "/etc/networkTables.gert"
-
-local function waitWithCancel(timeout, cancelCheck)
-	local now = computer.uptime()
-	local deadline = now + timeout
-	while now < deadline do
-		event.pull(deadline - now, "modem_message")
-		local response = cancelCheck()
-		if response then return response end
-		now = computer.uptime()
-	end
-	return cancelCheck()
-end
 
 local function storeChild(rAddress, receiveM, port, tier)
 	local childGA
