+++ GERTi/MNCInstaller.lua
@@ -1,0 +1,3 @@
+local shell = require("shell")
+shell.execute("wget https://raw.githubusercontent.com/GlobalEmpire/GERT/master/GERTi/GERTiClient.lua /etc/rc.d/GERTiMNC.lua")
+shell.execute("wget /lib/MNCAPI.lua")
